pyproject.toml
[project]
name = "evl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "rich>=14.3.3",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]
